Implement useAuthContext in Hike.jsx and HikeForm.jsx. Check to see if there is a user before allowing the creation of a new hike or the deletion of an existing hike. Also, create the UpdateHikeForm and the dispatch for the UPDATE_HIKE action type. But wating until authorization is introduced by the Ninja to implement the UpdateHike functinality into the app.

+++ src/components/Hike.jsx
@@ -1,18 +1,29 @@
 import { useHikesContext } from "../hooks/useHikesContext"
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
 import { Link } from "react-router-dom"
+import { useAuthContext } from "../hooks/useAuthContext"
 const Hike = ({ hike }) => {
   const { dispatch } = useHikesContext()
+  const { user } = useAuthContext()
   const { title, rating, image, description, createdAt } = hike
   console.log(image)
   const handleDelete = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch(
       `http://localhost:4000/api/hikes/${hike._id}`,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     )
+
     const deletedHike = await response.json()
+    console.log(deletedHike.error)
+
     if (response.ok) {
       dispatch({ type: "DELETE_HIKE", payload: deletedHike })
     }
